+++ changelog.md
@@ -2,3 +2,6 @@
 
 - Added a function to export mods into a JSON file (modpack.json)
 - Added a prototype logo
+- Added a button to export mods
+- Added a custom layer to view the button
+- Added a button to show the layer (right to the Geode button)